Move explain helpers into _explain_package
Move the effective explain logic and its helper functions out of
crawl.py and into _explain_package.py so the explain-specific code
lives together in one place.

As part of that extraction, move the shared sublime_text selector
parsing helpers into _utils.py so both crawl runtime logic and explain
logic use the same implementation.

Update the explain tests to import the helper-facing functions from
_explain_package.
